requirements.txt
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.0.6
django-crispy-forms==1.14.0
django-environ==0.9.0
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
